DISE/requirement.txt
pycryptodome
numpy
os